requirements.txt
discord.py>=2.3.2
openpyxl>=3.1.2
aiosqlite>=0.20.0
aiohttp>=3.9.5
python-dotenv>=1.0.1
